requirements.txt
# Core LLM & AI
openai==2.29.0
pydantic==2.12.5

# Web Framework
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1

# Database
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11

# Utilities
python-dotenv==1.2.2
